refactor: 도전과제 - 뱃지 리렌더링 최적화
